Refactor component structure to use index.js and style.scss Refactor styles to use BEM Add banner placeholde image
