--- src/components/Banner/Banner.scss
@@ -1,7 +1,0 @@
-.banner {
-  max-width: 100%;
-}
-
-.banner-image {
-  max-width: 100%;
-}
