+++ src/components/Banner/BannerImage/style.scss
@@ -1,0 +1,5 @@
+@import "~utils/base/base";
+
+.BannerImage {
+  max-width: 100%;
+}
