+++ src/components/Banner/style.scss
@@ -1,0 +1,5 @@
+@import "~utils/base/base";
+
+.Banner {
+  max-width: 100%;
+}
