+++ src/components/index.js
@@ -1,5 +1,5 @@
 import { actions } from 'sdk';
-import Banner from './Banner/Banner';
+import Banner from './Banner';
 
 let components = [
   {
