--- src/editors/BannerEditor/BannerEditor.scss
@@ -1,57 +1,0 @@
-@import "~utils/vtex-variables.scss";
-
-.v-banner-ed__editor {
-  background-color: $gray-soft;
-  height: 100%;
-  overflow: auto;
-}
-
-.v-banner-ed__editor__wrapper {
-  padding: 15px 15px 70px 15px;
-}
-
-.v-banner-ed__title {
-  font-family: $font-display;
-  font-weight: 400;
-  margin-top: 10px;
-}
-
-.v-banner-ed__add {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.v-banner-ed__add-label {
-  color: $black;
-  margin-top: 0.5em;
-  margin-left: 0.5em;
-  font-weight: 700;
-}
-
-.v-banner__show {
-  display: flex;
-}
-
-.v-banner__hide {
-  display: none;
-}
-
-.v-banner-ed__form__wrapper {
-  margin-top: 1.5em;
-}
-
-.v-banner-ed__form {
-  margin: 0em 1em 0em 1em;
-
-  input {
-    margin-bottom: 1em;
-  }
-}
-
-.v-banner-ed__form__url, .v-banner-ed__form__link, .v-banner-ed__form__alt {
-  label {
-    font-family: $font-body;
-    color: $black-soft;
-  }
-}
