+++ src/editors/BannerEditor/BannerPlaceholder/style.scss
@@ -1,0 +1,20 @@
+@import "~utils/base/base";
+
+.BannerPlaceholder {
+
+  align-items: center;
+  background: url('#{$asset-path}/images/placeholder.png') no-repeat $gray-solid;
+  background-position: center;
+  background-size: 100px auto;
+  border-radius: $pad / 2;
+  display: flex;
+  height: 190px;
+  justify-content: center;
+
+  &__text {
+    display: flex;
+    font-size: 1.2em;
+    font-weight: 700;
+    color: $black-soft;
+  }
+}
